indoxRouter_server/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1
httpx>=0.24.0
openai>=1.0.0
anthropic>=0.5.0
google-generativeai>=0.8.0
mistralai>=1.0.0
python-dotenv>=1.0.0
tiktoken>=0.5.0
aiohttp>=3.8.0
tenacity>=8.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0 